requirements.txt
Pillow==7.0.0
telethon==1.17.0
